Integrate edit view to main
Comment away the step sequencer logic sketch to make room for the
MVC driven UI.

+++ software/presentation/RhythmPatternEditing/inc/PatternEditView.h
@@ -31,8 +31,8 @@ private:
 
     inline void handleStateUpdate(ViewMode mode);
     inline void outputViewToLeds(ViewMode mode);
-    inline u8 calculateEditModeDrawBuffer();
-    inline u8 calculateSelectModeDrawBuffer();
+    inline u16 calculateEditModeDrawBuffer();
+    inline u16 calculateSelectModeDrawBuffer();
 };
 
 #endif /* PATTERN_EDIT_VIEW */
